[#3636] fetch current user, display in <TopBar> and set in userRdr
