+++ akvo/rsr/spa/app/store/root-reducer.js
@@ -1,9 +1,11 @@
 import { combineReducers } from 'redux'
 
 import editorRdr from '../modules/editor/reducer'
+import userRdr from './user-reducer'
 
 const rootReducer = combineReducers({
   editorRdr,
+  userRdr
 })
 
 export default rootReducer
