+++ akvo/rsr/spa/app/store/user-reducer.js
@@ -1,0 +1,9 @@
+const initialState = {}
+
+export default (state = initialState, action) => {
+  switch(action.type){
+    case 'SET_USER':
+      return action.user
+    default: return state
+  }
+}
